+++ project-base/src/Shopsys/ShopBundle/Resources/scripts/admin/categoryTree.formItem.js
@@ -8,9 +8,11 @@
         var self = this;
         self.STATUS_OPENED = 'opened';
         self.STATUS_CLOSED = 'closed';
+        self.STATUS_LOADING = 'loading';
         self.STATUS_NONE = 'none';
 
         var status = null;
+        var loaded = null;
         var $statusIcon = $item.find('.js-category-tree-form-item-icon:first');
         var $checkbox = $item.find('.js-category-tree-form-item-checkbox:first');
         self.parent = parent;
@@ -48,6 +50,9 @@
                     }
                 }
             }
+            if (loaded === null) {
+                loaded = self.children.length > 0;
+            }
         };
 
         this.statusToggle = function () {
@@ -59,13 +64,60 @@
         };
 
         this.open = function (animate) {
-            if (!$childrenContainer.is(':animated')) {
+            if (loaded === false) {
+                this.loadChildren();
+            } else if (!$childrenContainer.is(':animated')) {
                 $childrenContainer.slideDown(animate === true ? 'normal' : 0);
                 setStatus(self.STATUS_OPENED);
                 if (self.parent instanceof Shopsys.categoryTree.FormItem) {
                     self.parent.open(animate);
                 }
             }
+        };
+
+        this.loadChildren = function () {
+            setStatus(self.STATUS_LOADING);
+
+            Shopsys.ajax({
+                loaderElement: $item,
+                url: $item.data('load-url'),
+                dataType: 'json',
+                success: function (data) {
+                    loaded = true;
+
+                    $.each(data, function () {
+                        var $newItem = createItem(this);
+                        $childrenContainer.append($newItem);
+                    });
+                    initChildren();
+
+                    self.open(true);
+                },
+                complete: function () {
+                    if (status === self.STATUS_LOADING) {
+                        setStatus(self.STATUS_CLOSED);
+                    }
+                }
+            });
+        };
+
+        var createItem = function (itemData) {
+            var $form = $item.closest('.js-category-tree-form');
+            var newItemHtml = $form.data('prototype');
+
+            newItemHtml = newItemHtml.replace(/__name__/g, itemData.id);
+            newItemHtml = newItemHtml.replace(/__category_name__/g, itemData.categoryName);
+
+            var $newItem = $(newItemHtml);
+            $newItem.data('load-url', itemData.loadUrl);
+            $newItem.data('has-children', itemData.hasChildren);
+            if (itemData.isVisible === false) {
+                $newItem.addClass($form.data('hidden-item-class'));
+            }
+
+            $newItem.find('.js-category-tree-form-item-checkbox').val(itemData.id);
+
+            return $newItem;
         };
 
         this.close = function (animate) {
@@ -84,6 +136,7 @@
             $statusIcon.removeClass('svg svg-circle-plus svg-circle-remove sprite sprite-level cursor-pointer form-tree__item__icon--level');
             switch (status) {
                 case self.STATUS_OPENED:
+                case self.STATUS_LOADING:
                     $statusIcon.addClass('svg svg-circle-remove cursor-pointer');
                     break;
                 case self.STATUS_CLOSED: